1. Serialization/gs1/requirements.txt
absl-py==1.4.0
asgiref==3.7.2
astunparse==1.6.3
beautifulsoup4==4.12.2
bs4==0.0.1
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.3
cmake==3.27.0
colorama==0.4.6
contourpy==1.1.0
crispy-bootstrap5==0.7
cycler==0.11.0
distlib==0.3.7
Django==4.2.3
django-admin-honeypot==1.1.0
django-autoslug-field==0.2.3
django-cors-headers==4.2.0
django-crispy-forms==2.0
djangorestframework==3.14.0
dlib==19.24.2
EasyProcess==1.1
entrypoint2==1.1
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.42.0
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.57.0
h5py==3.9.0
idna==3.4
imutils==0.5.4
joblib==1.3.2
keras==2.13.1
kiwisolver==1.4.4
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
MouseInfo==0.1.3
mss==9.0.1
mysql-connector-python==8.1.0
numpy==1.24.3
oauthlib==3.2.2
opencv-contrib-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
platformdirs==3.9.1
protobuf==4.21.12
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pyqt5-tools==5.15.9.3.3
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytweening==1.0.7
pytz==2023.3
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
sslcommerz-lib==1.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.2.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
virtualenv==20.24.0
Werkzeug==2.3.7
wrapt==1.15.0
